consent manager bugfix, 404 page improvement
