+++ website/package.json
@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "author": "HashiCorp",
   "dependencies": {
-    "@hashicorp/nextjs-scripts": "^10.0.1",
+    "@hashicorp/nextjs-scripts": "^10.0.2",
     "@hashicorp/react-button": "^2.2.0",
     "@hashicorp/react-code-block": "^1.2.7",
     "@hashicorp/react-content": "3.0.0-0",
