+++ website/pages/not-found/index.jsx
@@ -1,0 +1,33 @@
+import s from './style.module.css'
+import Link from 'next/link'
+import { useEffect } from 'react'
+
+export default function NotFound() {
+  useEffect(() => {
+    if (
+      typeof window !== 'undefined' &&
+      typeof window?.analytics?.track === 'function' &&
+      typeof window?.document?.referrer === 'string' &&
+      typeof window?.location?.href === 'string'
+    )
+      window.analytics.track(window.location.href, {
+        category: '404 Response',
+        label: window.document.referrer || 'No Referrer',
+      })
+  }, [])
+
+  return (
+    <div className={s.root}>
+      <h1>Page Not Found</h1>
+      <p>
+        We&lsquo;re sorry but we can&lsquo;t find the page you&lsquo;re looking
+        for.
+      </p>
+      <p>
+        <Link href="/">
+          <a>Back to Home</a>
+        </Link>
+      </p>
+    </div>
+  )
+}
